# openapi.yml
openapi: 3.0.2
info:
  title: Structures service
  version: '0.1'
paths:
...
  /outlets/{outlet_id}:
    get:
      tags:
        - Outlets
        - Outlets
      summary: Get Outlet
      operationId: get_outlet_outlets__outlet_id__get
      parameters:
        - required: true
          schema:
            title: Outlet Id
            type: string
          name: outlet_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outlet'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    put:
      tags:
        - Outlets
        - Outlets
      summary: Update Outlet
      operationId: update_outlet_outlets__outlet_id__put
      parameters:
        - required: true
          schema:
            title: Outlet Id
            type: string
          name: outlet_id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutletUpdateDto'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outlet'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    delete:
      tags:
        - Outlets
        - Outlets
      summary: Delete Outlet
      operationId: delete_outlet_outlets__outlet_id__delete
      parameters:
        - required: true
          schema:
            title: Outlet Id
            type: string
          name: outlet_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /outlets:
    get:
      tags:
        - Outlets
        - Outlets
      summary: Find Outlets
      operationId: find_outlets_outlets_get
      parameters:
        - required: false
          schema:
            title: Active
            type: boolean
          name: active
          in: query
        - required: false
          schema:
            title: Child Of
            type: string
          name: child_of
          in: query
        - required: false
          schema:
            title: Page
            type: integer
            default: 1
          name: page
          in: query
        - required: false
          schema:
            title: Limit
            type: integer
            default: 100
          name: limit
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutletPaginated'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    post:
      tags:
        - Outlets
        - Outlets
      summary: Create Outlet
      operationId: create_outlet_outlets_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutletCreateDto'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outlet'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /outlets/{outlet_id}/activate:
    patch:
      tags:
        - Outlets
        - Outlets
      summary: Activate Outlet
      operationId: activate_outlet_outlets__outlet_id__activate_patch
      parameters:
        - required: true
          schema:
            title: Outlet Id
            type: string
          name: outlet_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outlet'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /outlets/{outlet_id}/deactivate:
    patch:
      tags:
        - Outlets
        - Outlets
      summary: Deactivate Outlet
      operationId: deactivate_outlet_outlets__outlet_id__deactivate_patch
      parameters:
        - required: true
          schema:
            title: Outlet Id
            type: string
          name: outlet_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outlet'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /outlets/{outlet_id}/change-parent/{new_parent_id}:
    patch:
      tags:
        - Outlets
        - Outlets
      summary: Change Parent Organization Unit
      operationId: change_parent_organization_unit_outlets__outlet_id__change_parent__new_parent_id__patch
      parameters:
        - required: true
          schema:
            title: Outlet Id
            type: string
          name: outlet_id
          in: path
        - required: true
          schema:
            title: New Parent Id
            type: string
          name: new_parent_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outlet'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
...
  /workers/{worker_id}:
    get:
      tags:
        - Workers
        - Workers
      summary: Get Worker
      operationId: get_worker_workers__worker_id__get
      parameters:
        - required: true
          schema:
            title: Worker Id
            type: string
          name: worker_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    put:
      tags:
        - Workers
        - Workers
      summary: Update Worker
      operationId: update_worker_workers__worker_id__put
      parameters:
        - required: true
          schema:
            title: Worker Id
            type: string
          name: worker_id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerUpdateDto'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    delete:
      tags:
        - Workers
        - Workers
      summary: Delete Worker
      operationId: delete_worker_workers__worker_id__delete
      parameters:
        - required: true
          schema:
            title: Worker Id
            type: string
          name: worker_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /workers:
    get:
      tags:
        - Workers
        - Workers
      summary: Find Workers
      operationId: find_workers_workers_get
      parameters:
        - required: false
          schema:
            title: Active
            type: boolean
          name: active
          in: query
        - required: false
          schema:
            title: Child Of Organization Unit
            type: string
          name: child_of_organization_unit
          in: query
        - required: false
          schema:
            title: Page
            type: integer
            default: 1
          name: page
          in: query
        - required: false
          schema:
            title: Limit
            type: integer
            default: 100
          name: limit
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerPaginatedDto'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
    post:
      tags:
        - Workers
        - Workers
      summary: Create Worker
      operationId: create_worker_workers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerCreateDto'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /workers/{worker_id}/activate:
    patch:
      tags:
        - Workers
        - Workers
      summary: Activate Worker
      operationId: activate_worker_workers__worker_id__activate_patch
      parameters:
        - required: true
          schema:
            title: Worker Id
            type: string
          name: worker_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /workers/{worker_id}/deactivate:
    patch:
      tags:
        - Workers
        - Workers
      summary: Deactivate Worker
      operationId: deactivate_worker_workers__worker_id__deactivate_patch
      parameters:
        - required: true
          schema:
            title: Worker Id
            type: string
          name: worker_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
  /workers/{worker_id}/change-organization/{new_parent_id}:
    patch:
      tags:
        - Workers
        - Workers
      summary: Change Organization Unit
      operationId: change_organization_unit_workers__worker_id__change_organization__new_parent_id__patch
      parameters:
        - required: true
          schema:
            title: Worker Id
            type: string
          name: worker_id
          in: path
        - required: true
          schema:
            title: New Parent Id
            type: string
          name: new_parent_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - UserIdSchema: []
...
components:
  schemas:
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    Outlet:
      title: Outlet
      required:
        - id
        - name
        - address
        - active
        - organization_unit_id
        - materialized_path
      type: object
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        address:
          title: Address
          type: string
        active:
          title: Active
          type: boolean
        organization_unit_id:
          title: Organization Unit Id
          type: string
        materialized_path:
          title: Materialized Path
          type: array
          items:
            type: string
    OutletBase:
      title: OutletBase
      required:
        - id
        - name
        - address
        - active
        - organization_unit_id
      type: object
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        address:
          title: Address
          type: string
        active:
          title: Active
          type: boolean
        organization_unit_id:
          title: Organization Unit Id
          type: string
    OutletCreateDto:
      title: OutletCreateDto
      required:
        - name
        - address
        - active
        - organization_unit_id
      type: object
      properties:
        name:
          title: Name
          type: string
        address:
          title: Address
          type: string
        active:
          title: Active
          type: boolean
        organization_unit_id:
          title: Organization Unit Id
          type: string
    OutletPaginated:
      title: OutletPaginated
      required:
        - pagination
        - data
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        data:
          title: Data
          type: array
          items:
            $ref: '#/components/schemas/OutletBase'
    OutletUpdateDto:
      title: OutletUpdateDto
      type: object
      properties:
        name:
          title: Name
          type: string
        address:
          title: Address
          type: string
    Pagination:
      title: Pagination
      required:
        - page
        - limit
        - count
      type: object
      properties:
        page:
          title: Page
          type: integer
        limit:
          title: Limit
          type: integer
        count:
          title: Count
          type: integer
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    Worker:
      title: Worker
      required:
        - id
        - fio
        - drivers_license
        - active
        - organization_unit_id
        - materialized_path
      type: object
      properties:
        id:
          title: Id
          type: string
        fio:
          title: Fio
          type: string
        drivers_license:
          title: Drivers License
          type: string
        active:
          title: Active
          type: boolean
        organization_unit_id:
          title: Organization Unit Id
          type: string
        materialized_path:
          title: Materialized Path
          type: array
          items:
            type: string
    WorkerCreateDto:
      title: WorkerCreateDto
      required:
        - fio
        - drivers_license
        - active
        - organization_unit_id
      type: object
      properties:
        fio:
          title: Fio
          type: string
        drivers_license:
          title: Drivers License
          type: string
        active:
          title: Active
          type: boolean
        organization_unit_id:
          title: Organization Unit Id
          type: string
    WorkerPaginatedDto:
      title: WorkerPaginatedDto
      required:
        - pagination
        - data
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        data:
          title: Data
          type: array
          items:
            $ref: '#/components/schemas/Worker'
    WorkerUpdateDto:
      title: WorkerUpdateDto
      type: object
      properties:
        fio:
          title: Fio
          type: string
        drivers_license:
          title: Drivers License
          type: string
  securitySchemes:
    UserIdSchema:
      type: apiKey
      in: header
      name: X-User-Id